pyproject.toml
[project]
name = "mkdocs-landing"
version = "0.0.8"
authors = [
    {name = "Todd Parsons", email = "todd.e.parsons@gmail.com"}
]
description = "mkdocs theme geared towards making a personal landing page, with simpler navigation and heavily customisable style options."
readme = "README.md"

dependencies = [
    "mkdocs"
]
[project.optional-dependencies]
tests = [
    "pytest",
    "httpx",
    "pyyaml",
    "mkdocs-get-deps",
    "mkdocs-material",
    "shot-scraper",
    "jinja2",
    "tqdm",
    "uv",
    "mkdocstrings-python",
]

[project.urls]
Homepage = "https://TEParsons.github.io/mkdocs-landing"
Download = "https://github.com/TEParsons/mkdocs-landing/releases"

[tool.setuptools]
packages = ["mkdocs_landing"]

[tool.setuptools.package-data]
mkdocs_landing = ["**/*.html", "**/*.css", "**/*.yml", "**/*.jpg", "**/*.png", "**/*.json", "**/*.ttf"]

[project.entry-points."mkdocs.themes"]
landing = "mkdocs_landing"

[project.entry-points."mkdocs.plugins"]
landing-preset = "mkdocs_landing.presets.plugin:LandingPresetsPlugin"
